feat: add first implementation to evaluate multiple polylogs via GiNaC
Added new mpl_ and lst_ functions for the following syntax of the multiple polylogs: mpl_(lst_(indices),lst_(args)). Added the procedure CalculateMpl that performs the evaluation using GiNaC. The arguments are parsed via new GetMplArgument procedure, to be merged with the existing GetLinArgument later. Input checks on the arguments of mpl_ and lst_ are still pending.